TASK#12: Creates hook useFilmLoad
