+++ project/src/screens/add-review/add-review.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from 'react';
+import React, { useEffect } from 'react';
 import { Link, useParams } from 'react-router-dom';
 import Header from '../../components/header/header';
 import { scrollToFilmTitle } from '../../utils/side-effects';
@@ -6,9 +6,8 @@ import ReviewForm from '../../components/review-form/review-form';
 import { State } from '../../types/state';
 import { connect, ConnectedProps } from 'react-redux';
 import NotFound from '../not-found/not-found';
-import { fetchFilm } from '../../services/dal';
-import { Film } from '../../types/film';
 import Spinner from '../../components/spinner/Spinner';
+import { useLoadFilm } from '../../hooks/films';
 
 const mapStateToProps = (state: State) => ({
   films: state.films,
@@ -20,27 +19,11 @@ type AddReviewProps = ConnectedProps<typeof connector>;
 
 function AddReview(props: AddReviewProps): JSX.Element {
   const { films } = props;
-  const [ isLoading, setIsLoading ] = useState(true);
-  const [ film, setFilm ] = useState<Film | null>(null);
 
   const { id } = useParams<{ id: string }>();
   useEffect(scrollToFilmTitle);
 
-  useEffect(() => {
-    const result = films.find((item) => +item.id === +id);
-
-    if (result) {
-      setFilm(result);
-      setIsLoading(false);
-      return;
-    }
-
-    fetchFilm(id)
-      .then((data) => {
-        setFilm(data);
-        setIsLoading(false);
-      });
-  }, [ films, id ]);
+  const { isLoading, film } = useLoadFilm(id, films);
 
   if (isLoading) {
     return (
